Implement Command Line Interface
Adding a CLI to play the game requires some small additions to the API to be able to, for example, show the state of the board on every player's movement. It also uncovers the requirement to know the mined neighbors of a coordinate.
These changes keep the initial modularized approach. Please notice how `cli` module is defined independently from the squared implementation and then everything is joined in `cli-squared` module.